+++ src/EntityGraphQL/Compiler/GraphQLExecutableNode.cs
@@ -4,6 +4,6 @@ namespace EntityGraphQL.Compiler
 {
     public abstract class GraphQLExecutableNode
     {
-        public abstract object Execute<TContext>(TContext context, IServiceProvider serviceProvider);
+        public abstract object Execute<TContext>(TContext context, GraphQLValidator validator, IServiceProvider serviceProvider);
     }
 }