+++ src/EntityGraphQL/Schema/SchemaProvider.cs
@@ -89,7 +89,6 @@ namespace EntityGraphQL.Schema
             QueryResult result;
             try
             {
-                GraphQLVaildation.Errors = new List<GraphQLError>(); //clear existing errors
                 var graphQLCompiler = new GraphQLCompiler(this, methodProvider);
                 var queryResult = graphQLCompiler.Compile(gql, claims);
                 result = queryResult.ExecuteQuery(context, serviceProvider, gql.OperationName);
@@ -164,7 +163,8 @@ namespace EntityGraphQL.Schema
                     string name = SchemaGenerator.ToCamelCaseStartsLower(method.Name);
                     var claims = method.GetCustomAttributes(typeof(GraphQLAuthorizeAttribute)).Cast<GraphQLAuthorizeAttribute>();
                     var requiredClaims = new RequiredClaims(claims);
-                    var mutationType = new MutationType(name, types[GetSchemaTypeNameForClrType(method.ReturnType)], mutationClassInstance, method, attribute.Description, requiredClaims);
+                    var typeName = GetSchemaTypeNameForClrType(method.ReturnType);
+                    var mutationType = new MutationType(name, types.ContainsKey(typeName) ? types[typeName] : null, mutationClassInstance, method, attribute.Description, requiredClaims);
                     mutations[name] = mutationType;
                 }
             }
@@ -620,20 +620,6 @@ namespace EntityGraphQL.Schema
             types.Add(name, schemaType);
             return schemaType.AddAllFields();
         }
-
-        /// <summary>
-        /// Add a graphql error
-        /// </summary>
-        /// <param name="message"></param>
-        public void AddError(string message)
-        {
-            GraphQLVaildation.Errors.Add(new GraphQLError(message));
-        }
-
-        /// <summary>
-        /// Check if any GraphQL vaildation error
-        /// </summary>
-        public bool IsVaild { get => GraphQLVaildation.Errors.Count == 0; }
 
         public IDirectiveProcessor GetDirective(string name)
         {